requirements.txt
Django==1.3
geopy
south==0.7.3
django-piston==0.2.2